requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
cryptography==41.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
aiofiles==23.2.1